fix: index version of path check, and fix existence query
